Made sure we only remove left over media that exists on the file system (this might  not be the case on first checkouts)

+++ apps/unisubs_compressor/management/commands/compile_media.py
@@ -206,7 +206,8 @@ class Command(BaseCommand):
         for filename in NO_UNIQUE_URL:
             to_path =  os.path.join(settings.MEDIA_ROOT, filename)
             from_path = os.path.join(final_path, filename)
-            os.remove(to_path)
+            if os.path.exsits(to_path):
+                os.remove(to_path)
             shutil.copy(from_path, to_path)
